fix: replace 'vim' hook with generic 'edit' hook
