+++ sodalite/core/sodalite.conf
@@ -69,13 +69,13 @@ hooks:
       action: 'sodalite-open "$entry" 2> /dev/null'
       label: open
   plain_text:
-    v:
-      action: 'vim "$entry"'
-      label: vim
-  executable:
     e:
+      action: '${EDITOR:-$(which vim || which vi || which nano || echo NO_EDITOR_FOUND)} "$entry"'
+      label: edit
+  executable:
+    r:
       action: 'exec "$entry"'
-      label: execute
+      label: run
   custom:
     archive:
       extensions: [xz]
