+++ sodalite/ui/notify.py
@@ -5,8 +5,9 @@ from typing import Union, Tuple, Optional
 import urwid
 from urwid import AttrSpec
 
-from sodalite.ui import graphics, theme, viewmodel
-from sodalite.ui.viewmodel import Mode
+import sodalite.ui.mmode
+from sodalite.ui import theme, mmode
+from sodalite.ui.mmode import Mode
 from sodalite.util import pubsub
 
 txt = urwid.AttrMap(urwid.Text('', align='center'), urwid.DEFAULT)
@@ -26,9 +27,19 @@ def show(message: Union[str, Tuple[AttrSpec, str]], duration: float = 1.5) -> No
     thread.start()
 
 
+def show_error(message: str, duration: float = 1.5) -> None:
+    """
+    Triggers a notification
+    :param message: the error to display
+    :param duration: if set to 0, the message is displayed until another call to 'show' or 'clear'
+    """
+    show((AttrSpec(theme.forbidden + ',bold', '', colors=16), message), duration)
+
+
 def _show(message: Union[str, Tuple[AttrSpec, str]], duration: float) -> None:
     global _last_message
     global _original_footer
+    from sodalite.ui import graphics
     frame = graphics.frame
     if _notify_lock.locked() and _last_message == message:
         return
@@ -48,15 +59,16 @@ def _show(message: Union[str, Tuple[AttrSpec, str]], duration: float) -> None:
 
 
 def clear() -> None:
+    from sodalite.ui import graphics
     if graphics.frame and graphics.frame.footer == _notify_box:
         graphics.frame.footer = None
         graphics.draw_screen()
 
 
 def trigger_notifications(mode: Mode) -> None:
-    if viewmodel.global_mode == Mode.ASSIGN_CHOOSE_ENTRY:
+    if mmode.global_mode == Mode.ASSIGN_CHOOSE_ENTRY:
         show("choose entry", duration=0)
-    elif viewmodel.global_mode == Mode.ASSIGN_CHOOSE_KEY:
+    elif sodalite.ui.mmode.global_mode == Mode.ASSIGN_CHOOSE_KEY:
         show("choose new key", duration=0)
     else:
         clear()
